Add support for qemu extended key events
This lets us send keycodes instead of just keysyms. The input message is
modified to take an optional 3rd argument indicating the raw XT Scancode
of the key that was pressed.

If the vnc server supports the qemu extended key extension, then we'll
send this scancode/keycode along with the keysym to the server. This
helps work around some keyboard oddities/inconsistencies between keysyms
and keycodes.